+++ src/framework/piral-core/src/types/plugin.ts
@@ -2,7 +2,11 @@ import { PiletApi } from './api';
 import { GlobalStateContext } from './state';
 import { PiletMetadata } from '../types';
 
-export interface ApiExtender<T> {
+/**
+ * Shape to be used by a Pilet API extension that requires other
+ * APIs or some metadata to work properly.
+ */
+export interface PiletApiExtender<T> {
   /**
    * Extends the base API of a module with new functionality.
    * @param api The API created by the base layer.
@@ -12,11 +16,14 @@ export interface ApiExtender<T> {
   (api: PiletApi, target: PiletMetadata): T;
 }
 
-export interface Extend<T = Partial<PiletApi>> {
+/**
+ * Shape to be used by a plugin used in the Piral instance.
+ */
+export interface PiralPlugin<T = Partial<PiletApi>> {
   /**
    * Extends the base API with a custom set of functionality to be used by modules.
    * @param context The global state context to be used.
    * @returns The extended API or a function to create the extended API for a specific target.
    */
-  (context: GlobalStateContext): T | ApiExtender<T>;
+  (context: GlobalStateContext): T | PiletApiExtender<T>;
 }