+++ src/pages/docs/src/codegen/markdown.js
@@ -7,6 +7,7 @@ const markdownItEmoji = require('markdown-it-emoji');
 const markdownItFootnote = require('markdown-it-footnote');
 const markdownItFrontMatter = require('markdown-it-front-matter');
 const markdownItHljs = require('markdown-it-highlightjs');
+const markdownItContainer = require('markdown-it-container');
 const markdownItMark = require('markdown-it-mark');
 const markdownItReplaceLink = require('markdown-it-replace-link');
 const markdownItSmartArrows = require('markdown-it-smartarrows');
@@ -14,9 +15,10 @@ const markdownItSub = require('markdown-it-sub');
 const markdownItSup = require('markdown-it-sup');
 const markdownItVideo = require('markdown-it-video');
 const { readFileSync } = require('fs');
-const { extname, basename, relative } = require('path');
+const { extname, basename } = require('path');
 const { createHash } = require('crypto');
 const { docRef, imgRef, rootPath } = require('./utils');
+const { makeRelativePath } = require('./paths');
 
 function computeHash(content) {
   return createHash('sha1')
@@ -36,6 +38,42 @@ function getMdValue(result) {
       .split(id).join('${require("' + path + '")}');
   });
   return content;
+}
+
+function wrapContainer(containerName, utils) {
+  const prefix = `${containerName}:`;
+  return {
+    validate(params) {
+      return params.trim().startsWith(prefix);
+    },
+    render(tokens, idx) {
+      if (tokens[idx].nesting === 1) {
+        const titleHtml = tokens[idx].info.trim().substring(prefix.length).trim();
+        const title = utils.escapeHtml(titleHtml);
+        return `<div class="box ${containerName}"><p class="box-title">${title}</p>\n`;
+      } else {
+        return '</div>\n';
+      }
+    },
+  };
+}
+
+function wrapCollapsed(containerName, utils) {
+  const prefix = `${containerName}:`;
+  return {
+    validate(params) {
+      return params.trim().startsWith(prefix);
+    },
+    render(tokens, idx) {
+      if (tokens[idx].nesting === 1) {
+        const titleHtml = tokens[idx].info.trim().substring(prefix.length).trim();
+        const title = utils.escapeHtml(titleHtml);
+        return `<details class="${containerName}"><summary>${title}</summary>\n`;
+      } else {
+        return '</details>\n';
+      }
+    },
+  };
 }
 
 function render(file, baseDir = __dirname) {
@@ -59,7 +97,7 @@ function render(file, baseDir = __dirname) {
         const content = readFileSync(target);
         const hash = computeHash(content);
         const id = `${name}_${hash}${ext}`;
-        result.images[id] = relative(baseDir, target);
+        result.images[id] = makeRelativePath(baseDir, target);
         return id;
       } else if (/LICENSE$/.test(link)) {
         return docRef(link, rootPath);
@@ -67,21 +105,29 @@ function render(file, baseDir = __dirname) {
 
       return link;
     },
-  })
-    .use(markdownItAbbr)
-    .use(markdownItAnchor, { level: [1, 2] })
+  });
+
+  md.use(markdownItAbbr)
+    .use(markdownItAnchor, { level: [1, 2, 3, 4, 5, 6] })
     .use(markdownItEmoji)
     .use(markdownItFootnote)
     .use(markdownItFrontMatter, fm => (result.meta = YAML.parse(fm)))
     .use(markdownItHljs)
+    .use(markdownItContainer, 'warning', wrapContainer('warning', md.utils))
+    .use(markdownItContainer, 'tip', wrapContainer('tip', md.utils))
+    .use(markdownItContainer, 'failure', wrapContainer('failure', md.utils))
+    .use(markdownItContainer, 'question', wrapContainer('question', md.utils))
+    .use(markdownItContainer, 'success', wrapContainer('success', md.utils))
+    .use(markdownItContainer, 'summary', wrapCollapsed('summary', md.utils))
     .use(markdownItMark)
     .use(markdownItReplaceLink)
     .use(markdownItSmartArrows)
     .use(markdownItSub)
     .use(markdownItSup)
     .use(markdownItVideo);
+
   result.content = md.render(content);
-  result.mdValue = '`' + getMdValue(result) + '`';
+  result.mdValue = ['`', getMdValue(result), '`'].join('');
   return result;
 }